requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
pep8==1.7.1
python-dotenv==0.10.3
toml==0.10.0
Werkzeug==0.15.4
